Implement all unit tests
